requirements.txt
python-telegram-bot==20.8
yfinance==0.2.36
pandas>=1.3.0
numpy>=1.21.0
requests>=2.25.0
python-dotenv
